:truck: Move teardown of context resources into 'shm teardown'
