+++ data_safe_haven/commands/context.py
@@ -9,9 +9,7 @@ from data_safe_haven.context import (
     Context,
     ContextSettings,
 )
-from data_safe_haven.context_infrastructure import ContextInfrastructure
 from data_safe_haven.exceptions import (
-    DataSafeHavenAzureAPIAuthenticationError,
     DataSafeHavenConfigError,
 )
 from data_safe_haven.logging import get_logger
@@ -200,31 +198,3 @@ def remove(
         raise typer.Exit(code=1) from exc
     settings.remove(key)
     settings.write()
-
-
-@context_command_group.command()
-def teardown() -> None:
-    """Tear down Data Safe Haven context infrastructure."""
-    logger = get_logger()
-    try:
-        context = ContextSettings.from_file().assert_context()
-    except DataSafeHavenConfigError as exc:
-        if exc.args[0] == "No context selected":
-            logger.critical(
-                "No context selected. Use `dsh context switch` to select one."
-            )
-        else:
-            logger.critical(
-                "No context configuration file. Use `dsh context add` before creating infrastructure."
-            )
-        raise typer.Exit(code=1) from exc
-
-    context_infra = ContextInfrastructure(context)
-
-    try:
-        context_infra.teardown()
-    except DataSafeHavenAzureAPIAuthenticationError as exc:
-        logger.critical(
-            "Failed to authenticate with the Azure API. You may not be logged into the Azure CLI, or your login may have expired. Try running `az login`."
-        )
-        raise typer.Exit(1) from exc
