+++ data_safe_haven/commands/shm.py
@@ -121,28 +121,52 @@ def deploy(
 @shm_command_group.command()
 def teardown() -> None:
     """Tear down a deployed a Safe Haven Management environment."""
-    context = ContextSettings.from_file().assert_context()
-    config = SHMConfig.from_remote(context)
-    pulumi_config = DSHPulumiConfig.from_remote(context)
+    # Teardown Data Safe Haven context infrastructure.
+    logger = get_logger()
+    try:
+        context = ContextSettings.from_file().assert_context()
+    except DataSafeHavenConfigError as exc:
+        if exc.args[0] == "No context selected":
+            logger.critical(
+                "No context selected. Use `dsh context switch` to select one."
+            )
+        else:
+            logger.critical(
+                "No context configuration file. Use `dsh context add` before creating infrastructure."
+            )
+        raise typer.Exit(code=1) from exc
 
     try:
-        # Remove infrastructure deployed with Pulumi
+        context_infra = ContextInfrastructure(context)
+        context_infra.teardown()
+    except DataSafeHavenAzureAPIAuthenticationError as exc:
+        logger.critical(
+            "Failed to authenticate with the Azure API. You may not be logged into the Azure CLI, or your login may have expired. Try running `az login`."
+        )
+        raise typer.Exit(1) from exc
+
+    # Teardown Data Safe Haven SHM infrastructure deployed with Pulumi
+    config = SHMConfig.from_remote(context)
+    pulumi_config = DSHPulumiConfig.from_remote(context)
+    try:
         try:
+            # Teardown the Pulumi project
             stack = SHMProjectManager(
                 context=context,
                 config=config,
                 pulumi_config=pulumi_config,
             )
             stack.teardown()
+
+            # Remove information from config file
+            del pulumi_config[context.shm_name]
+
+            # Upload Pulumi config to blob storage
+            pulumi_config.upload(context)
         except Exception as exc:
             msg = "Unable to teardown Pulumi infrastructure."
             raise DataSafeHavenInputError(msg) from exc
 
-        # Remove information from config file
-        del pulumi_config[context.shm_name]
-
-        # Upload Pulumi config to blob storage
-        pulumi_config.upload(context)
     except DataSafeHavenError as exc:
         msg = "Could not teardown Safe Haven Management environment."
         raise DataSafeHavenError(msg) from exc
