+++ tests/commands/test_context.py
@@ -1,7 +1,4 @@
 from data_safe_haven.commands.context import context_command_group
-from data_safe_haven.context_infrastructure import ContextInfrastructure
-from data_safe_haven.exceptions import DataSafeHavenAzureAPIAuthenticationError
-from data_safe_haven.external.interface.azure_authenticator import AzureAuthenticator
 
 
 class TestShow:
@@ -222,36 +219,3 @@ class TestRemove:
         )
         assert result.exit_code == 1
         assert "No context configuration file." in result.stdout
-
-
-class TestTeardown:
-    def test_teardown(self, runner, monkeypatch):
-        def mock_teardown(self):  # noqa: ARG001
-            print("mock teardown")  # noqa: T201
-
-        monkeypatch.setattr(ContextInfrastructure, "teardown", mock_teardown)
-
-        result = runner.invoke(context_command_group, ["teardown"])
-        assert "mock teardown" in result.stdout
-        assert result.exit_code == 0
-
-    def test_no_context_file(self, runner_no_context_file):
-        result = runner_no_context_file.invoke(context_command_group, ["teardown"])
-        assert result.exit_code == 1
-        assert "No context configuration file." in result.stdout
-
-    def test_show_none(self, runner_none):
-        result = runner_none.invoke(context_command_group, ["teardown"])
-        assert result.exit_code == 1
-        assert "No context selected." in result.stdout
-
-    def test_auth_failure(self, runner, mocker):
-        def mock_login(self):  # noqa: ARG001
-            msg = "Failed to authenticate with Azure API."
-            raise DataSafeHavenAzureAPIAuthenticationError(msg)
-
-        mocker.patch.object(AzureAuthenticator, "login", mock_login)
-
-        result = runner.invoke(context_command_group, ["teardown"])
-        assert result.exit_code == 1
-        assert "Failed to authenticate with the Azure API." in result.stdout
